qr_scanner_backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gunicorn>=21.2
whitenoise>=6.5
psycopg2-binary>=2.9
dj-database-url>=2.1
qrcode[pil]>=7.4
